Create a base class for response cache
