+++ src/cache/response.ts
@@ -1,179 +1,2 @@
-import { AdapterDependencies } from '../adapter'
-import { AdapterSettings } from '../config'
-import {
-  AdapterResponse,
-  ResponseGenerics,
-  TimestampedAdapterResponse,
-  TimestampedProviderErrorResponse,
-  TimestampedProviderResult,
-  censor,
-  censorLogs,
-  makeLogger,
-} from '../util'
-import CensorList from '../util/censor/censor-list'
-import {
-  InputParameters,
-  InputParametersDefinition,
-  TypeFromDefinition,
-} from '../validation/input-params'
-import { validator } from '../validation/utils'
-import { Cache, calculateCacheKey, calculateFeedId, calculateAdapterName } from './'
-import * as cacheMetrics from './metrics'
-
-const logger = makeLogger('ResponseCache')
-
-/**
- * Special type of cache to store responses for this adapter.
- */
-export class ResponseCache<
-  T extends {
-    Parameters: InputParametersDefinition
-    Response: ResponseGenerics
-  },
-> {
-  cache: Cache<AdapterResponse<T['Response']>>
-  inputParameters: InputParameters<T['Parameters']>
-  adapterName: string
-  endpointName: string
-  adapterSettings: AdapterSettings
-
-  constructor({
-    inputParameters,
-    adapterName,
-    endpointName,
-    adapterSettings,
-    dependencies,
-  }: {
-    dependencies: AdapterDependencies
-    adapterSettings: AdapterSettings
-    adapterName: string
-    endpointName: string
-    inputParameters: InputParameters<T['Parameters']>
-  }) {
-    this.cache = dependencies.cache as Cache<AdapterResponse<T['Response']>>
-    this.inputParameters = inputParameters
-    this.adapterName = adapterName
-    this.endpointName = endpointName
-    this.adapterSettings = adapterSettings
-  }
-
-  /**
-   * Sets responses in the adapter cache (adding necessary metadata and defaults)
-   *
-   * @param results - the entries to write to the cache
-   */
-  async write(transportName: string, results: TimestampedProviderResult<T>[]): Promise<void> {
-    const censorList = CensorList.getAll()
-    const entries = results.map((r) => {
-      const { data, result, errorMessage } = r.response
-      if (!errorMessage && data === undefined) {
-        logger.warn('The "data" property of the response is undefined.')
-      } else if (!errorMessage && result === undefined) {
-        logger.warn('The "result" property of the response is undefined.')
-      }
-      let censoredResponse
-      if (!censorList.length) {
-        censoredResponse = r.response
-      } else {
-        try {
-          censoredResponse = censor(r.response, censorList, true) as TimestampedAdapterResponse<
-            T['Response']
-          >
-        } catch (error) {
-          censorLogs(() => logger.error(`Error censoring response: ${error}`))
-          censoredResponse = {
-            statusCode: 502,
-            errorMessage: 'Response could not be censored due to an error',
-            timestamps: r.response.timestamps,
-          }
-        }
-      }
-
-      const response: AdapterResponse<T['Response']> = {
-        ...censoredResponse,
-        statusCode: (censoredResponse as TimestampedProviderErrorResponse).statusCode || 200,
-      }
-
-      if (
-        this.adapterSettings.METRICS_ENABLED &&
-        this.adapterSettings.EXPERIMENTAL_METRICS_ENABLED
-      ) {
-        response.meta = {
-          adapterName: calculateAdapterName(this.adapterName, r.params),
-          metrics: {
-            feedId: calculateFeedId(
-              {
-                adapterSettings: this.adapterSettings,
-              },
-              r.params,
-            ),
-          },
-        }
-      }
-
-      if (response.timestamps?.providerIndicatedTimeUnixMs !== undefined) {
-        const timestampValidator = validator.responseTimestamp()
-        const error = timestampValidator.fn(response.timestamps?.providerIndicatedTimeUnixMs)
-        if (error) {
-          censorLogs(() => logger.warn(`Provider indicated time is invalid: ${error}`))
-        }
-      }
-
-      return {
-        key: calculateCacheKey({
-          transportName,
-          data: r.params,
-          adapterName: this.adapterName,
-          endpointName: this.endpointName,
-          adapterSettings: this.adapterSettings,
-        }),
-        value: response,
-      } as const
-    })
-
-    const ttl = this.adapterSettings.CACHE_MAX_AGE
-    await this.cache.setMany(entries, ttl)
-
-    const now = Date.now()
-    for (const { key, value } of entries) {
-      // Only record metrics if feed Id is present, otherwise assuming value is not adapter response to record
-      const response = value as unknown as AdapterResponse
-      const feedId = response.meta?.metrics?.feedId
-      if (feedId) {
-        const providerTime = response.timestamps?.providerIndicatedTimeUnixMs
-        const timeDelta = providerTime ? now - providerTime : undefined
-
-        // Record cache set count, max age, and staleness (set to 0 for cache set)
-        const label = cacheMetrics.cacheMetricsLabel(key, feedId, this.cache.type)
-        cacheMetrics.cacheSet(label, ttl, timeDelta)
-      }
-    }
-
-    return
-  }
-
-  /**
-   * Sets a new TTL value for already cached responses in the adapter cache
-   *
-   * @param transportName - transport name
-   * @param params - set of parameters that uniquely relate to the response
-   * @param ttl - a new time in milliseconds until the response expires
-   */
-  async writeTTL(
-    transportName: string,
-    params: TypeFromDefinition<T['Parameters']>[],
-    ttl: number,
-  ): Promise<void> {
-    for (const param of params) {
-      const key = calculateCacheKey({
-        transportName: transportName,
-        data: param,
-        adapterName: this.adapterName,
-        endpointName: this.endpointName,
-        adapterSettings: this.adapterSettings,
-      })
-
-      this.cache.setTTL(key, ttl)
-    }
-  }
-}
+export { ResponseCache } from './response-cache/base'
+export { SimpleResponseCache } from './response-cache/simple'
